refactor: remove unit tests from integration test file
- Remove direct testing of private stream methods
- Add unit test to system_info module where it can access internals
- Remove test-support feature that was added but not needed
- Follow the pattern of other platforms (Linux, Windows) that only test via full dump
