+++ tests/ios_minidump_writer.rs
@@ -343,8 +343,6 @@ mod macos_tests {
         MinidumpSystemInfo, MinidumpThreadList, MinidumpThreadNames, Module,
     };
     use minidump_common::format::PlatformId;
-    use minidump_writer::apple::ios::task_dumper::TaskDumper;
-    use minidump_writer::dir_section::DumpBuf;
     use minidump_writer::ios_test::*;
     use minidump_writer::minidump_format::*;
     use scroll::Pread;
@@ -359,25 +357,6 @@ mod macos_tests {
         let mut cursor = Cursor::new(Vec::new());
         MinidumpWriter::new().dump(&mut cursor)?;
         Ok(cursor.into_inner())
-    }
-
-    #[test]
-    fn test_write_system_info() {
-        let mut writer = MinidumpWriter::new();
-        let dumper = TaskDumper::new(unsafe { mach2::traps::mach_task_self() }).unwrap();
-        let mut buffer = DumpBuf::with_capacity(0);
-
-        // Write system info
-        let result = writer.write_system_info(&mut buffer, &dumper);
-        assert!(result.is_ok());
-
-        let dirent = result.unwrap();
-        assert_eq!(dirent.stream_type, MDStreamType::SystemInfoStream as u32);
-        assert!(dirent.location.data_size > 0);
-        assert_eq!(
-            dirent.location.data_size as usize,
-            std::mem::size_of::<MDRawSystemInfo>()
-        );
     }
 
     #[test]
@@ -427,12 +406,12 @@ mod macos_tests {
 
     #[test]
     fn test_system_info_contents() {
-        let mut writer = MinidumpWriter::new();
+        let mut writer = TestMinidumpWriter::new();
         let dumper = TaskDumper::new(unsafe { mach2::traps::mach_task_self() }).unwrap();
         let mut buffer = DumpBuf::with_capacity(0);
 
         // Write system info
-        let result = writer.write_system_info(&mut buffer, &dumper);
+        let result = writer.test_write_system_info(&mut buffer, &dumper);
         assert!(result.is_ok());
 
         // Read back the system info
@@ -482,7 +461,7 @@ mod macos_tests {
 
     #[test]
     fn test_minidump_writer_with_system_info() {
-        let mut writer = MinidumpWriter::new();
+        let mut writer = TestMinidumpWriter::new();
         let mut cursor = Cursor::new(Vec::new());
 
         // Dump to cursor
@@ -527,13 +506,13 @@ mod macos_tests {
 
     #[test]
     fn test_thread_list_direct() {
-        let mut writer = MinidumpWriter::new();
+        let mut writer = TestMinidumpWriter::new();
         let task = unsafe { mach2::traps::mach_task_self() };
         let dumper = TaskDumper::new(task).unwrap();
         let mut buffer = DumpBuf::with_capacity(0);
 
         // Write thread list directly
-        let result = writer.write_thread_list(&mut buffer, &dumper);
+        let result = writer.test_write_thread_list(&mut buffer, &dumper);
         assert!(result.is_ok());
 
         let (dirent, _) = result.unwrap();
@@ -580,7 +559,7 @@ mod macos_tests {
 
     #[test]
     fn test_thread_list_stream() {
-        let mut writer = MinidumpWriter::new();
+        let mut writer = TestMinidumpWriter::new();
         let mut cursor = Cursor::new(Vec::new());
 
         // Dump full minidump to get proper thread list
@@ -750,14 +729,14 @@ mod macos_tests {
 
     #[test]
     fn test_stack_overflow_handling() {
-        let mut writer = MinidumpWriter::new();
+        let mut writer = TestMinidumpWriter::new();
         let task = unsafe { mach2::traps::mach_task_self() };
         let dumper = TaskDumper::new(task).unwrap();
         let mut buffer = DumpBuf::with_capacity(0);
 
         // We can't easily simulate a real stack overflow, but we can test
         // the handling logic by checking that the sentinel values are properly used
-        let result = writer.write_thread_list(&mut buffer, &dumper);
+        let result = writer.test_write_thread_list(&mut buffer, &dumper);
         assert!(result.is_ok());
 
         let (dirent, _) = result.unwrap();
@@ -832,7 +811,7 @@ mod macos_tests {
 
     #[test]
     fn test_crashed_thread_with_context() {
-        let mut writer = MinidumpWriter::new();
+        let mut writer = TestMinidumpWriter::new();
         let task = unsafe { mach2::traps::mach_task_self() };
         let current_thread = unsafe { mach2::mach_init::mach_thread_self() };
 
@@ -850,13 +829,13 @@ mod macos_tests {
         };
 
         // Set the crash context on the writer
-        writer = MinidumpWriter::with_crash_context(crash_context);
+        writer = TestMinidumpWriter::with_crash_context(crash_context);
 
         let dumper = TaskDumper::new(task).unwrap();
         let mut buffer = DumpBuf::with_capacity(0);
 
         // Write thread list with crash context
-        let result = writer.write_thread_list(&mut buffer, &dumper);
+        let result = writer.test_write_thread_list(&mut buffer, &dumper);
         assert!(result.is_ok());
 
         let (_dirent, crashed_thread_context) = result.unwrap();
@@ -875,13 +854,13 @@ mod macos_tests {
 
     #[test]
     fn test_memory_list_stream() {
-        let mut writer = MinidumpWriter::new();
+        let mut writer = TestMinidumpWriter::new();
         let task = unsafe { mach2::traps::mach_task_self() };
         let dumper = TaskDumper::new(task).unwrap();
         let mut buffer = DumpBuf::with_capacity(0);
 
         // First write thread list to populate memory_blocks
-        let result = writer.write_thread_list(&mut buffer, &dumper);
+        let result = writer.test_write_thread_list(&mut buffer, &dumper);
         assert!(result.is_ok());
 
         // Verify we have some memory blocks from thread stacks
@@ -895,7 +874,7 @@ mod macos_tests {
         let initial_blocks = 1; // Assume at least one block
 
         // Now write memory list
-        let memory_result = writer.write_memory_list(&mut buffer, &dumper);
+        let memory_result = writer.test_write_memory_list(&mut buffer, &dumper);
         assert!(memory_result.is_ok());
 
         let dirent = memory_result.unwrap();
@@ -921,7 +900,7 @@ mod macos_tests {
 
     #[test]
     fn test_memory_list_with_exception() {
-        let mut writer = MinidumpWriter::new();
+        let mut writer = TestMinidumpWriter::new();
         let task = unsafe { mach2::traps::mach_task_self() };
         let current_thread = unsafe { mach2::mach_init::mach_thread_self() };
 
@@ -943,17 +922,17 @@ mod macos_tests {
         };
 
         // Set the crash context on the writer
-        writer = MinidumpWriter::with_crash_context(crash_context);
+        writer = TestMinidumpWriter::with_crash_context(crash_context);
 
         let mut buffer = DumpBuf::with_capacity(0);
 
         // Write thread list first
-        writer.write_thread_list(&mut buffer, &dumper).unwrap();
+        writer.test_write_thread_list(&mut buffer, &dumper).unwrap();
         // Can't access private field in tests
         // let blocks_before = writer.memory_blocks.len();
 
         // Write memory list - should include IP memory for exception
-        let result = writer.write_memory_list(&mut buffer, &dumper);
+        let result = writer.test_write_memory_list(&mut buffer, &dumper);
         assert!(result.is_ok());
 
         // With an exception, we might have added memory around the IP
@@ -1118,7 +1097,7 @@ mod macos_tests {
 
     #[test]
     fn test_module_list_stream() {
-        let mut writer = MinidumpWriter::new();
+        let mut writer = TestMinidumpWriter::new();
         let mut cursor = Cursor::new(Vec::new());
 
         // Dump full minidump to get module list
@@ -1199,7 +1178,7 @@ mod macos_tests {
 
     #[test]
     fn test_thread_register_capture() {
-        let mut writer = MinidumpWriter::new();
+        let mut writer = TestMinidumpWriter::new();
         let mut cursor = Cursor::new(Vec::new());
 
         // Dump full minidump
@@ -1274,7 +1253,7 @@ mod macos_tests {
 
     #[test]
     fn test_breakpad_info_stream() {
-        let mut writer = MinidumpWriter::new();
+        let mut writer = TestMinidumpWriter::new();
         let mut cursor = Cursor::new(Vec::new());
 
         // Dump full minidump
@@ -1296,7 +1275,7 @@ mod macos_tests {
 
     #[test]
     fn test_thread_names_stream() {
-        let mut writer = MinidumpWriter::new();
+        let mut writer = TestMinidumpWriter::new();
         let mut cursor = Cursor::new(Vec::new());
 
         // Dump full minidump
@@ -1333,7 +1312,7 @@ mod macos_tests {
 
     #[test]
     fn test_misc_info_stream() {
-        let mut writer = MinidumpWriter::new();
+        let mut writer = TestMinidumpWriter::new();
         let mut cursor = Cursor::new(Vec::new());
 
         // Dump full minidump
@@ -1362,7 +1341,7 @@ mod macos_tests {
 
     #[test]
     fn test_breakpad_info_with_crash_context() {
-        let mut writer = MinidumpWriter::new();
+        let mut writer = TestMinidumpWriter::new();
         let task = unsafe { mach2::traps::mach_task_self() };
         let current_thread = unsafe { mach2::mach_init::mach_thread_self() };
 
@@ -1380,7 +1359,7 @@ mod macos_tests {
         };
 
         // Set the crash context on the writer
-        writer = MinidumpWriter::with_crash_context(crash_context);
+        writer = TestMinidumpWriter::with_crash_context(crash_context);
 
         let mut cursor = Cursor::new(Vec::new());
 
@@ -1445,7 +1424,7 @@ mod macos_tests {
 
     #[test]
     fn test_stream_count_with_new_streams() {
-        let mut writer = MinidumpWriter::new();
+        let mut writer = TestMinidumpWriter::new();
         let mut cursor = Cursor::new(Vec::new());
 
         // Dump full minidump without exception
@@ -1489,7 +1468,7 @@ mod macos_tests {
     #[test]
     fn test_module_base_address_calculation() {
         // Test for fix in commit a743db0c: module base address should be (vm_addr + slide)
-        let mut writer = MinidumpWriter::new();
+        let mut writer = TestMinidumpWriter::new();
         let mut cursor = Cursor::new(Vec::new());
 
         // Dump full minidump
@@ -1621,7 +1600,7 @@ mod macos_tests {
     #[test]
     fn test_thread_register_values_in_minidump() {
         // Test that thread register values are properly captured in minidump
-        let mut writer = MinidumpWriter::new();
+        let mut writer = TestMinidumpWriter::new();
         let mut cursor = Cursor::new(Vec::new());
 
         // Dump full minidump
